+++ src/core/time/timer.impl.hpp
@@ -21,8 +21,8 @@ public:
     void start();
     void snapshot();
 
-    uint elapsedMilliseconds() const;
-    uint effectiveElapsedMilliseconds() const;
+    float elapsedMilliseconds() const;
+    float effectiveElapsedMilliseconds() const;
 
     int fps() const;
 
@@ -37,16 +37,16 @@ private:
     void delay(uint ms) const;
     uint ticks() const;
 
-    uint m_start {0};
-    uint m_elapsedMilliseconds {0};
-    uint m_effectiveElapsedMilliseconds {0};
+    float m_start {0};
+    float m_elapsedMilliseconds {0};
+    float m_effectiveElapsedMilliseconds {0};
 
-    uint m_fpsElapsedMilliseconds {0};
+    float m_fpsElapsedMilliseconds {0};
     uint m_fpsTemp {0};
     uint m_fps {0};
 
     uint m_fixedFps {0};
-    uint m_fixedFpsDelayTime {0};
+    float m_fixedFpsDelayTime {0};
 
     uint m_totalFrames {0};
 };
